+++ enterprise_access/apps/content_assignments/api.py
@@ -12,7 +12,6 @@ from opaque_keys import InvalidKeyError
 from opaque_keys.edx.keys import CourseKey
 from opaque_keys.edx.locator import CourseLocator
 
-from enterprise_access.apps.content_metadata.api import get_and_cache_catalog_content_metadata
 from enterprise_access.apps.core.models import User
 from enterprise_access.apps.subsidy_access_policy.content_metadata_api import get_and_cache_content_metadata
 
@@ -451,28 +450,6 @@ def cancel_assignments(assignments: Iterable[LearnerContentAssignment]) -> dict:
     return {
         'cancelled': list(set(cancelled_assignments) | already_cancelled_assignments),
         'non_cancelable': list(non_cancelable_assignments),
-    }
-
-
-def get_content_metadata_for_assignments(enterprise_catalog_uuid, assignments):
-    """
-    Fetches (from cache or enterprise-catalog API call) content metadata
-    in bulk for the `content_keys` of the given assignments, provided
-    such metadata is related to the given `enterprise_catalog_uuid`.
-
-    Returns:
-        A dict mapping every content key of the provided assignments
-        to a content metadata dictionary, or null if no such dictionary
-        could be found for a given key.
-    """
-    content_keys = sorted({assignment.content_key for assignment in assignments})
-    content_metadata_list = get_and_cache_catalog_content_metadata(enterprise_catalog_uuid, content_keys)
-    metadata_by_key = {
-        record['key']: record for record in content_metadata_list
-    }
-    return {
-        assignment.content_key: metadata_by_key.get(assignment.content_key)
-        for assignment in assignments
     }
